refactor: remove ohter card layouts. - consolidate to use only OauthCardLayout.tsx - remove unused card layouts
- refactor: rename OauthCardLayout to AuthCardLayout
- makes the layout more generic and reusable
